+++ classes/Visualizer/Gutenberg/Block.php
@@ -46,7 +46,7 @@ class Visualizer_Gutenberg_Block {
 	 * Returns an instance of this class.
 	 */
 	public static function get_instance() {
-		if ( null == self::$instance ) {
+		if ( null === self::$instance ) {
 			self::$instance = new Visualizer_Gutenberg_Block();
 		}
 		return self::$instance;
@@ -86,7 +86,7 @@ class Visualizer_Gutenberg_Block {
 		if ( VISUALIZER_PRO ) {
 			$type = 'pro';
 			if ( apply_filters( 'visualizer_is_business', false ) ) {
-				$type = 'business';
+				$type = 'developer';
 			}
 		}
 
@@ -121,11 +121,13 @@ class Visualizer_Gutenberg_Block {
 	 * Gutenberg Block Callback Function
 	 */
 	public function gutenberg_block_callback( $attr ) {
-		$id = $attr['id'];
-		if ( empty( $id ) || $id === 'none' ) {
-			return ''; // no id = no fun
+		if ( isset( $attr['id'] ) ) {
+			$id = $attr['id'];
+			if ( empty( $id ) || $id === 'none' ) {
+				return ''; // no id = no fun
+			}
+			return '[visualizer id="' . $id . '"]';
 		}
-		return '[visualizer id="' . $id . '"]';
 	}
 
 	/**
@@ -295,19 +297,19 @@ class Visualizer_Gutenberg_Block {
 				continue;
 			}
 
-			if ( $row['type'] == 'number' ) {
+			if ( $row['type'] === 'number' ) {
 				foreach ( $data as $o => $col ) {
 					$data[ $o ][ $i ] = ( is_numeric( $col[ $i ] ) ) ? floatval( $col[ $i ] ) : ( is_numeric( str_replace( ',', '', $col[ $i ] ) ) ? floatval( str_replace( ',', '', $col[ $i ] ) ) : null );
 				}
 			}
 
-			if ( $row['type'] == 'boolean' ) {
+			if ( $row['type'] === 'boolean' ) {
 				foreach ( $data as $o => $col ) {
 					$data[ $o ][ $i ] = ! empty( $col[ $i ] ) ? filter_validate( $col[ $i ], FILTER_VALIDATE_BOOLEAN ) : null;
 				}
 			}
 
-			if ( $row['type'] == 'timeofday' ) {
+			if ( $row['type'] === 'timeofday' ) {
 				foreach ( $data as $o => $col ) {
 					$date = new DateTime( '1984-03-16T' . $col[ $i ] );
 					if ( $date ) {
@@ -321,7 +323,7 @@ class Visualizer_Gutenberg_Block {
 				}
 			}
 
-			if ( $row['type'] == 'string' ) {
+			if ( $row['type'] === 'string' ) {
 				foreach ( $data as $o => $col ) {
 					$data[ $o ][ $i ] = $this->toUTF8( $col[ $i ] );
 				}