+++ js/render-facade.js
@@ -11,10 +11,13 @@
                 window.alert(v.i10n['copied']);
             });
         }
-        $('a.visualizer-action[data-visualizer-type!=copy]').on('click', function(e) {
+        $('a.visualizer-action[data-visualizer-type!=copy]').off('click').on('click', function(e) {
             var type    = $(this).attr( 'data-visualizer-type' );
             var chart   = $(this).attr( 'data-visualizer-chart-id' );
+            var container   = $(this).attr( 'data-visualizer-container-id' );
             var lock    = $('.visualizer-front.visualizer-front-' + chart);
+            var mime    = $(this).attr( 'data-visualizer-mime' );
+            console.log(mime);
             lock.lock();
             e.preventDefault();
             $.ajax({
@@ -23,38 +26,37 @@
                     if (data && data.data) {
                         switch(type){
                             case 'csv':
-                                var a = document.createElement("a");
-                                document.body.appendChild(a);
-                                a.style = "display: none";
-                                var blob = new Blob([data.data.csv], {type: $(this).attr( 'data-visualizer-mime' ) }),
-                                    url = window.URL.createObjectURL(blob);
-                                a.href = url;
-                                a.download = data.data.name;
-                                a.click();
-                                setTimeout(function () {
-                                    window.URL.revokeObjectURL(url);
-                                }, 100);
+                                var blob = new Blob([data.data.csv], {type: mime });
+                                if(window.navigator.msSaveOrOpenBlob){
+                                    window.navigator.msSaveOrOpenBlob(blob, data.data.name);
+                                } else {
+                                    var url = window.URL.createObjectURL(blob);
+                                    var $a = $("<a>");
+                                    $a.attr("href", url);
+                                    $("body").append($a);
+                                    $a.attr("download", data.data.name);
+                                    $a[0].click();
+                                    setTimeout(function () {
+                                        window.URL.revokeObjectURL(url);
+                                        $a.remove();
+                                    }, 100);
+                                }
                                 break;
                             case 'xls':
-                                var $a = $("<a>");
-                                $a.attr("href",data.data.csv);
-                                $("body").append($a);
-                                $a.attr("download",data.data.name);
-                                $a[0].click();
-                                $a.remove();
+                                if(window.navigator.msSaveOrOpenBlob){
+                                    blob = new Blob([s2ab(atob(data.data.raw))], {type: '' });
+                                    window.navigator.msSaveOrOpenBlob(blob, data.data.name);
+                                } else {
+                                    var $a = $("<a>"); // jshint ignore:line
+                                    $a.attr("href", data.data.csv);
+                                    $("body").append($a);
+                                    $a.attr("download", data.data.name);
+                                    $a[0].click();
+                                    $a.remove();
+                                }
                                 break;
                             case 'print':
-                                var iframe = $('<iframe>').attr("name", "print-visualization").attr("id", "print-visualization").css("position", "absolute");
-                                iframe.appendTo($('body'));
-                                var iframe_doc = iframe.get(0).contentWindow || iframe.get(0).contentDocument.document || iframe.get(0).contentDocument;
-                                iframe_doc.document.open();
-                                iframe_doc.document.write(data.data.csv);
-                                iframe_doc.document.close();
-                                setTimeout(function(){
-                                    window.frames['print-visualization'].focus();
-                                    window.frames['print-visualization'].print();
-                                    iframe.remove();
-                                }, 500);
+                                $('body').trigger('visualizer:action:specificchart', {action: 'print', id: container, data: data.data.csv});
                                 break;
                             default:
                                 if(window.visualizer_perform_action) {
@@ -69,10 +71,36 @@
         });
     }
 
+    function s2ab(s) {
+        var buf = new ArrayBuffer(s.length);
+        var view = new Uint8Array(buf);
+        for (var i=0; i !== s.length; ++i) {
+            view[i] = s.charCodeAt(i) & 0xFF;
+        }
+        return buf;
+    }
+
     $(document).ready(function(){
         $('body').trigger('visualizer:render:chart:start', visualizer);
         initActionsButtons(visualizer);
+        registerDefaultActions();
     });
+
+    function registerDefaultActions(){
+        $('body').off('visualizer:action:specificchart:defaultprint').on('visualizer:action:specificchart:defaultprint', function(event, v){
+            var iframe = $('<iframe>').attr("name", "print-visualization").attr("id", "print-visualization").css("position", "absolute");
+            iframe.appendTo($('body'));
+            var iframe_doc = iframe.get(0).contentWindow || iframe.get(0).contentDocument.document || iframe.get(0).contentDocument;
+            iframe_doc.document.open();
+            iframe_doc.document.write(v.data);
+            iframe_doc.document.close();
+            setTimeout(function(){
+                window.frames['print-visualization'].focus();
+                window.frames['print-visualization'].print();
+                iframe.remove();
+            }, 500);
+        });
+    }
 })(jQuery, visualizer);
 
 (function ($) {