+++ CHANGELOG.md
@@ -1,4 +1,29 @@
 
+ ### v3.2.0 - 2019-05-03 
+ **Changes:** 
+ * Add support for charts in AMP requests
+* Add support to show charts from JSON/REST endpoints
+* Fix loading of Google Visualization javascript files
+* Add simple editors for editing chart data
+* Tested up to WP 5.2
+ 
+ ### v3.1.3 - 2019-02-24 
+ **Changes:** 
+ * Fix issue with changing column settings of the last column in table chart
+* Add support for query language to get subset of data from Google Spreadsheet
+* Fix conflict with jquery 3.3.x
+* Migrated PHPExcel to PhpSpreadsheet
+* Front end action 'print' should print the chart and fall back to printing the data
+* Fix issue with table chart not showing in IE
+* Fix issue with multiple instances of same chart not showing
+* Fix issue with date type column does not work with Combo charts
+* Tested with WP 5.1
+ 
+ ### v3.1.2 - 2018-12-06 
+ **Changes:** 
+ * Fix bug "Warning: A non-numeric value encountered"
+* Tested with WP 5.0
+ 
  ### v3.1.1 - 2018-12-05 
  **Changes:** 
  * Fix issue with Gutenberg support